2a: Split render.go into render.go + render_subagent.go
